auto make crd manifests
